+++ README.md
@@ -1,6 +1,8 @@
 # BetaBoost
 A small wrapper to do Beta Boosting with XgBoost
 
+```pip install BetaBoost==0.0.4```
+
 Initiate a BetaBoost object and fit a XgBoost model.
 Returns a XgBoost Train Object.
 
@@ -13,6 +15,7 @@ import numpy as np
 import xgboost as xgb
 import matplotlib.pyplot as plt
 
+xgb.__version__
 
 def generate_data():
     y = np.random.gamma(2, 4, OBS)
@@ -48,7 +51,7 @@ model1 = xgb.train(
     evals=[(dtrain, 'train'),(dtest, 'test')],
     evals_result=progress1,
     verbose_eval=False,
-    learning_rates=eta_decay
+    callbacks=[xgb.callback.LearningRateScheduler(eta_decay)]
 )
 
 progress2 = dict()
@@ -61,7 +64,7 @@ model2 = xgb.train(
     evals=[(dtrain, 'train'),(dtest, 'test')],
     evals_result=progress2,
     verbose_eval=False,
-    learning_rates=list(np.ones(max_iter)*0.01)
+    callbacks=[xgb.callback.LearningRateScheduler(list(np.ones(max_iter)*0.01))]
 )
 
 
@@ -75,7 +78,7 @@ model3 = xgb.train(
     evals=[(dtrain, 'train'),(dtest, 'test')],
     evals_result=progress3,
     verbose_eval=False,
-    learning_rates=list(np.ones(max_iter)*0.1)
+    callbacks=[xgb.callback.LearningRateScheduler(list(np.ones(max_iter)*0.1))]
 )
 
 #Here we call the BetaBoost, the wrapper parameters are passed in the class init
@@ -83,10 +86,12 @@ bb_evals = dict()
 import BetaBoost as bb
 betabooster = bb.BetaBoost(n_boosting_rounds = max_iter)
 betabooster.fit(dtrain=dtrain,
-          params = PARAMS,
-          evals_result = bb_evals,
-          verbose_eval = False,
-          dtest = dtest)
+                maximize=True,
+                params = PARAMS,
+                early_stopping_rounds=max_iter,
+                evals=[(dtrain, 'train'),(dtest, 'test')],
+                evals_result = bb_evals,
+                verbose_eval = False)
 
 plt.plot(progress1['test']['rmse'], linestyle = 'dashed', color = 'b', label = 'eta test decay')
 plt.plot(progress2['test']['rmse'], linestyle = 'dashed', color = 'r', label = '0.01 test')
@@ -103,7 +108,7 @@ A quick convergence and robustness to overfitting can be achieved with slight tu
 
 FAQ:
 
-Why a beta kernel?
+Why a beta pdf?
 
 TLDR: it worked pretty well!
